Repository setup
find_restaurant intent is complete. To get this intent. All the utterances of the user (until one with location is found) are concatenated to form a single one. This brings some problems, since there will be no find_restaurant training samples without the location entity, being hard after to rasa to assume this intent without a location present.